requirements.txt
pytest-playwright==0.4.3
playwright-stealth==1.0.6
oauth2client~=4.1.3
pyotp==2.9.0
gspread
openai
gdown
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
Pillow
python-dotenv
typer
sqlalchemy
pymysql==1.1.1
tenacity
dependency-injector==4.43.0
mysql-connector-python==8.0.33